fix(parity): f64 print_z for layer templates; silent zero-length deretract
Upstream accumulates print_z in double precision (Print layer print_z);
the schedule's f32 truncation made template guards like
layer_z >= initial_layer_print_height + layer_height * 2 fail on ULP
boundaries (Raise3D Pro3's before-layer-change M106 P2 ramp emitted S100
where orca emits S150 at exactly z=0.7). Layer z's now flow as f64
through the schedule/boundary chain; the HEIGHT header keeps the
f32-difference quirk (upstream's ;HEIGHT:0.200001 artifacts).

GCodeWriter::unretract emits the deretract G1 only for a non-zero dE:
retraction_length=0 printers (re3D GigabotX, wipe-only retraction) no
longer emit a spurious G1 E0 F1800.

re3D GigabotX 2 (all nozzles) and Raise3D Pro3 (all variants) now match
byte-for-byte.

+++ crates/ares-core/src/project_slice/gcode_emit/layers.rs
@@ -258,7 +258,10 @@ pub(super) fn append(
             0.0
         };
         let layer_z = group_z;
-        let layer_height = layer_z - previous_layer_z;
+        // The HEIGHT header keeps the f32-difference quirk (upstream
+        // `;HEIGHT:0.200001` artifacts) while layer_z itself stays the
+        // f64 print_z for template comparisons.
+        let layer_height = f64::from(layer_z as f32) - f64::from(previous_layer_z as f32);
         let boundary = boundary::append(
             output,
             state,
